+++ src/Vortice.DirectML/ActivationGeluOperatorDescription.cs
@@ -1,0 +1,55 @@
+// Copyright © Aaron Sun, Amer Koleci, and Contributors.
+// Licensed under the MIT License (MIT). See LICENSE in the repository root for more information.
+
+namespace Vortice.DirectML;
+
+public partial struct ActivationGeluOperatorDescription : IOperatorDescription, IOperatorDescriptionMarshal
+{
+    /// <summary>
+    /// Gets the type of operator description.
+    /// </summary>
+    public OperatorType OperatorType => OperatorType.ActivationGelu;
+
+    /// <include file="Documentation.xml" path="/comments/comment[@id='DML_ACTIVATION_GELU_OPERATOR_DESC::InputTensor']/*" />
+    public TensorDescription InputTensor { get; set; }
+
+    /// <include file="Documentation.xml" path="/comments/comment[@id='DML_ACTIVATION_GELU_OPERATOR_DESC::OutputTensor']/*" />
+    public TensorDescription OutputTensor { get; set; }
+
+    /// <inheritdoc></inheritdoc>/>
+    public override string ToString() => $"ActivationGelu";
+
+    #region Marshal
+    [StructLayout(LayoutKind.Sequential, Pack = 0)]
+    internal struct __Native
+    {
+        public IntPtr InputTensor;
+        public IntPtr OutputTensor;
+    }
+
+    unsafe IntPtr IOperatorDescriptionMarshal.__MarshalAlloc()
+    {
+        __Native* @ref = UnsafeUtilities.Alloc<__Native>();
+
+        @ref->InputTensor = InputTensor.__MarshalAlloc();
+        @ref->OutputTensor = OutputTensor.__MarshalAlloc();
+
+        return new(@ref);
+    }
+
+    unsafe void IOperatorDescriptionMarshal.__MarshalFree(ref IntPtr pDesc)
+    {
+        var @ref = (__Native*)pDesc;
+
+        InputTensor.__MarshalFree(ref @ref->InputTensor);
+        OutputTensor.__MarshalFree(ref @ref->OutputTensor);
+
+        UnsafeUtilities.Free(@ref);
+    }
+    #endregion
+
+    public static implicit operator OperatorDescription(ActivationGeluOperatorDescription description)
+    {
+        return new(description);
+    }
+}